Add HTTPS support to fakeintake client
